functions/requirements.txt
firebase_functions~=0.1.0
firebase-admin>=6.0.0
boto3>=1.34.0
requests>=2.31.0
google-cloud-translate>=3.12.0